add fluentValidation and update commands

+++ src/Domain/Commands/Handlers/ProductCommandHandler.cs
@@ -3,12 +3,14 @@ using Domain.Commands.Responses;
 using Domain.Contracts.Commands;
 using Domain.Contracts.Repositories;
 using Domain.Entities;
+using Domain.Validations;
 
 namespace Domain.Commands.Handlers
 {
-    public class ProductCommandHandler : Notifiable, IProductCommandHandler
+    public class ProductCommandHandler : IProductCommandHandler
     {
         private readonly IProductRepository _repository;
+
         public ProductCommandHandler(IProductRepository repository)
         {
             _repository = repository;
@@ -16,16 +18,24 @@ namespace Domain.Commands.Handlers
 
         public async Task<GenericResponse> Handle(ProductCreateCommand command)
         {
-            var exist = await _repository.Exists(command.Name);
-            if (exist)
-                return new GenericResponse(false, "Já existe um Produto cadastrado com esse Nome. ", command);
+            if (command is null)
+            {
+                return new GenericResponse(false, "Product is invalid!");
+            }
+
+            var validationResult = command.Validate();
+
+            if (!validationResult.IsValid)
+            {
+                return new GenericResponse(false, ErrorParser.GetErrorMessage(validationResult));
+            }
+
+            if (await _repository.Exists(command.Name))
+            {
+                return new GenericResponse(false, "Já existe um Produto cadastrado com esse Nome. ", command.Name);
+            }
 
             var product = new Product(command.Name, command.Price);
-
-            AddNotifications(product);
-
-            if (Invalid)
-                return new GenericResponse(false, string.Join(". ", Notifications));
 
             await _repository.Add(product);
 
@@ -34,20 +44,31 @@ namespace Domain.Commands.Handlers
 
         public async Task<GenericResponse> Handle(ProductUpdateCommand command)
         {
-            var product = await _repository.Get(command.Id);
-            if (product == null)
-                return new GenericResponse(false, "Produto não encontrado na base de dados. ", null);
+            if (command is null)
+            {
+                return new GenericResponse(false, "Product is invalid!");
+            }
 
-            var existOtherProduct = await _repository.ExistsUpdate(command.Name, command.Id);
-            if (existOtherProduct)
-                return new GenericResponse(false, "Já existe outro Produto cadatrado com esse Nome. ", null);
+            var validationResult = command.Validate();
+
+            if (!validationResult.IsValid)
+            {
+                return new GenericResponse(false, ErrorParser.GetErrorMessage(validationResult));
+            }
+
+            if (await _repository.ExistsUpdate(command.Name, command.Id))
+            {
+                return new GenericResponse(false, "Já existe um Produto cadastrado com esse Nome. ", command.Name);
+            }
+
+            var product = await _repository.Get(command.Id);
+
+            if (product is null)
+            {
+                return new GenericResponse(false, "Produto não encontrado na base de dados. ");
+            }
 
             product.Update(command.Name, command.Price);
-
-            AddNotifications(product);
-
-            if (Invalid)
-                return new GenericResponse(false, string.Join(". ", Notifications));
 
             await _repository.Update(product);
 
@@ -56,16 +77,26 @@ namespace Domain.Commands.Handlers
 
         public async Task<GenericResponse> Handle(ProductPromotionCommand command)
         {
+            if (command is null)
+            {
+                return new GenericResponse(false, "Product is invalid!");
+            }
+
+            var validationResult = command.Validate();
+
+            if (!validationResult.IsValid)
+            {
+                return new GenericResponse(false, ErrorParser.GetErrorMessage(validationResult));
+            }
+
             var product = await _repository.Get(command.Id);
+
             if (product == null)
-                return new GenericResponse(false, "Produto não encontrado na base de dados. ", null);
+            {
+                return new GenericResponse(false, "Produto não encontrado na base de dados. ");
+            }
 
             product.AddPromotion(command.Price);
-
-            AddNotifications(product);
-
-            if (Invalid)
-                return new GenericResponse(false, string.Join(". ", Notifications));
 
             await _repository.Update(product);
 
@@ -74,9 +105,17 @@ namespace Domain.Commands.Handlers
 
         public async Task<GenericResponse> Handle(ProductDeleteCommand command)
         {
+            if (command is null)
+            {
+                return new GenericResponse(false, "Product is invalid!");
+            }
+
             var product = await _repository.Get(command.Id);
+
             if (product == null)
-                return new GenericResponse(false, "Produto não encontrado na base de dados. ", null);
+            {
+                return new GenericResponse(false, "Produto não encontrado na base de dados. ");
+            }
 
             await _repository.Delete(product.Id);
